feat(applicationsettings): added cosmetic changes required to make errors look more understandable and added missing changes and helper file to make code more readable
